@@ -19,7 +19,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -29,16 +28,17 @@ import (
 	"strings"
 	"time"
 
-	// Preload proto definitions
-	_ "github.com/cncf/udpa/go/udpa/type/v1"
+	_ "github.com/cncf/udpa/go/udpa/type/v1" // Preload proto definitions
 	bootstrap_v3 "github.com/envoyproxy/go-control-plane/envoy/config/bootstrap/v3"
 	core "github.com/envoyproxy/go-control-plane/envoy/config/core/v3"
-
-	// Preload proto definitions
-	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/router/v3"
+	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/http/router/v3" // Preload proto definitions
 	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/http_inspector/v3"
 	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/listener/tls_inspector/v3"
 	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/http_connection_manager/v3"
+	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/filters/network/tcp_proxy/v3"
+	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/internal_upstream/v3"
+	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/transport_sockets/raw_buffer/v3"
+	_ "github.com/envoyproxy/go-control-plane/envoy/extensions/upstreams/http/v3"
 
 	"istio.io/proxy/test/envoye2e/env"
 )
@@ -87,7 +87,7 @@ func (e *Envoy) Run(p *Params) error {
 	}
 	log.Printf("admin port %d", e.adminPort)
 
-	tmp, err := ioutil.TempFile(os.TempDir(), "envoy-bootstrap-*.yaml")
+	tmp, err := os.CreateTemp(os.TempDir(), "envoy-bootstrap-*.yaml")
 	if err != nil {
 		return err
 	}
@@ -209,7 +209,7 @@ func downloadEnvoy(ver string) (string, error) {
 		return "", fmt.Errorf("cannot get envoy sha from %v: %v", proxyDepURL, err)
 	}
 	defer resp.Body.Close()
-	istioDeps, err := ioutil.ReadAll(resp.Body)
+	istioDeps, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return "", fmt.Errorf("cannot read body of istio deps: %v", err)
 	}
