@@ -100,9 +100,10 @@ const ::Wasm::Common::FlatNode& peerNodeInfo(
       fbb.GetBufferPointer());
 }
 
-::Wasm::Common::RequestInfo requestInfo() {
+::Wasm::Common::RequestInfo requestInfo(int response_code = 200) {
   ::Wasm::Common::RequestInfo request_info;
   request_info.start_time = 0;
+  request_info.response_code = response_code;
   request_info.request_operation = "GET";
   request_info.destination_service_host = "httpbin.org";
   request_info.destination_service_name = "httpbin";
@@ -162,7 +163,8 @@ std::string write_audit_request_json = R"({
            "referer":"www.google.com",
            "serverIp":"2.2.2.2",
            "latency":"10s",
-           "protocol":"http"
+           "protocol":"http",
+           "status":"200"
         },
         "timestamp":"1970-01-01T00:00:00Z",
         "severity":"INFO",
@@ -211,7 +213,8 @@ std::string write_log_request_json = R"({
            "referer":"www.google.com",
            "serverIp":"2.2.2.2",
            "latency":"10s",
-           "protocol":"http"
+           "protocol":"http",
+           "status":"200"
         },
         "timestamp":"1970-01-01T00:00:00Z",
         "severity":"INFO",
@@ -243,13 +246,76 @@ std::string write_log_request_json = R"({
   ]
 })";
 
+std::string write_error_log_request_json = R"({
+  "logName":"projects/test_project/logs/server-accesslog-stackdriver",
+  "resource":{
+     "type":"k8s_container",
+     "labels":{
+        "cluster_name":"test_cluster",
+        "pod_name":"test_pod",
+        "location":"test_location",
+        "namespace_name":"test_namespace",
+        "project_id":"test_project",
+        "container_name":"istio-proxy"
+     }
+  },
+  "labels":{
+     "destination_workload":"test_workload",
+     "mesh_uid":"mesh",
+     "destination_namespace":"test_namespace",
+     "destination_name":"test_pod"
+  },
+  "entries":[
+     {
+        "httpRequest":{
+           "requestMethod":"GET",
+           "requestUrl":"http://httpbin.org/headers?retry=true",
+           "userAgent":"chrome",
+           "remoteIp":"1.1.1.1",
+           "referer":"www.google.com",
+           "serverIp":"2.2.2.2",
+           "latency":"10s",
+           "protocol":"http",
+           "status":"404",
+        },
+        "timestamp":"1970-01-01T00:00:00Z",
+        "severity":"ERROR",
+        "labels":{
+           "source_name":"test_peer_pod",
+           "destination_principal":"destination_principal",
+           "destination_service_host":"httpbin.org",
+           "destination_service_name":"httpbin",
+           "request_id":"123",
+           "source_namespace":"test_peer_namespace",
+           "source_principal":"source_principal",
+           "service_authentication_policy":"MUTUAL_TLS",
+           "source_workload":"test_peer_workload",
+           "response_flag":"-",
+           "protocol":"http",
+           "log_sampled":"false",
+           "connection_id":"0",
+           "upstream_cluster": "server-inbound-cluster",
+           "route_name": "redirect",
+           "requested_server_name": "server.com",
+           "x-envoy-original-dst-host": "tmp.com",
+           "x-envoy-original-path": "/tmp",
+           "upstream_host": "1.1.1.1:1000"
+        },
+        "trace":"projects/test_project/traces/123abc",
+        "spanId":"abc123",
+        "traceSampled":true
+     }
+  ]
+})";
+
 google::logging::v2::WriteLogEntriesRequest expectedRequest(
-    int log_entry_count, bool for_audit = false) {
+    int log_entry_count, bool for_audit = false, bool use_error_log = false) {
   google::logging::v2::WriteLogEntriesRequest req;
   google::protobuf::util::JsonParseOptions options;
-  JsonStringToMessage(
-      (for_audit ? write_audit_request_json : write_log_request_json), &req,
-      options);
+  std::string non_audit_log =
+      use_error_log ? write_error_log_request_json : write_log_request_json;
+  JsonStringToMessage((for_audit ? write_audit_request_json : non_audit_log),
+                      &req, options);
   for (int i = 1; i < log_entry_count; i++) {
     auto* new_entry = req.mutable_entries()->Add();
     new_entry->CopyFrom(req.entries()[0]);
@@ -278,6 +344,34 @@ TEST(LoggerTest, TestWriteLogEntry) {
               MessageDifferencer differ;
               differ.ReportDifferencesToString(&diff);
               if (!differ.Compare(expectedRequest(1), *req)) {
+                FAIL() << "unexpected log entry " << diff << "\n";
+              }
+            }
+          }));
+  logger->exportLogEntry(/* is_on_done = */ false);
+}
+
+TEST(LoggerTest, TestWriteErrorLogEntry) {
+  auto exporter = std::make_unique<::testing::NiceMock<MockExporter>>();
+  auto exporter_ptr = exporter.get();
+  flatbuffers::FlatBufferBuilder local, peer;
+  std::unordered_map<std::string, std::string> extra_labels;
+  auto logger = std::make_unique<Logger>(nodeInfo(local), std::move(exporter),
+                                         extra_labels);
+  logger->addLogEntry(requestInfo(404), peerNodeInfo(peer), extra_labels, false,
+                      false);
+  EXPECT_CALL(*exporter_ptr, exportLogs(::testing::_, ::testing::_))
+      .WillOnce(::testing::Invoke(
+          [](const std::vector<std::unique_ptr<
+                 const google::logging::v2::WriteLogEntriesRequest>>& requests,
+             bool) {
+            for (const auto& req : requests) {
+              std::string diff;
+              MessageDifferencer differ;
+              differ.ReportDifferencesToString(&diff);
+              if (!differ.Compare(expectedRequest(1, false /* audit log */,
+                                                  true /* error log */),
+                                  *req)) {
                 FAIL() << "unexpected log entry " << diff << "\n";
               }
             }
