@@ -25,7 +25,8 @@ import (
 	"time"
 
 	"github.com/golang/protobuf/jsonpb"
-	"github.com/golang/protobuf/proto"
+	// nolint: staticcheck
+	legacyproto "github.com/golang/protobuf/proto"
 	yamlv2 "gopkg.in/yaml.v2"
 	"sigs.k8s.io/yaml"
 
@@ -150,6 +151,9 @@ func (p *Params) Fill(s string) (string, error) {
 				}
 				return out
 			},
+			"divisible": func(i, j int) bool {
+				return i%j == 0
+			},
 		}).
 		Parse(s))
 	var b bytes.Buffer
@@ -202,11 +206,11 @@ func (s *Scenario) Run(p *Params) error {
 
 func (s *Scenario) Cleanup() {}
 
-func ReadYAML(input string, pb proto.Message) error {
+func ReadYAML(input string, pb legacyproto.Message) error {
 	var jsonObj interface{}
 	err := yamlv2.Unmarshal([]byte(input), &jsonObj)
 	if err != nil {
-		return err
+		return fmt.Errorf("cannot parse: %v for %q", err, input)
 	}
 	// As a special case, convert [x] to x.
 	// This is needed because jsonpb is unable to parse arrays.
@@ -219,22 +223,21 @@ func ReadYAML(input string, pb proto.Message) error {
 	}
 	yml, err := yamlv2.Marshal(jsonObj)
 	if err != nil {
-		return err
+		return fmt.Errorf("cannot marshal: %v for %q", err, input)
 	}
 	js, err := yaml.YAMLToJSON(yml)
 	if err != nil {
-		return err
+		return fmt.Errorf("cannot unmarshal: %v for %q", err, input)
 	}
 	reader := strings.NewReader(string(js))
 	m := jsonpb.Unmarshaler{}
 	return m.Unmarshal(reader, pb)
 }
 
-func (p *Params) FillYAML(input string, pb proto.Message) error {
+func (p *Params) FillYAML(input string, pb legacyproto.Message) error {
 	out, err := p.Fill(input)
 	if err != nil {
 		return err
 	}
-	fmt.Println(out)
 	return ReadYAML(out, pb)
 }
