@@ -24,59 +24,84 @@ func init() {
 	// TODO(bianpengyuan): automatically generate this.
 	ProxyE2ETests = &env.TestInventory{
 		Tests: []string{
+			"TestAttributeGen/envoy.wasm.runtime.null",
+			"TestAttributeGen/envoy.wasm.runtime.v8",
 			"TestBasicFlow",
 			"TestBasicHTTP",
-			"TestBasicHTTPwithTLS",
 			"TestBasicHTTPGateway",
+			"TestBasicHTTPwithTLS",
 			"TestBasicTCPFlow",
+			"TestBasicCONNECT",
 			"TestHTTPExchange",
-			"TestStackdriverAccessLog/StackdriverAndAccessLogPlugin",
-			"TestStackdriverAccessLog/RequestGetsLoggedAgain",
-			"TestStackdriverAccessLog/AllErrorRequestsGetsLogged",
+			"TestHTTPLocalRatelimit",
 			"TestStackdriverAccessLog/AllClientErrorRequestsGetsLoggedOnNoMxAndError",
+			"TestStackdriverAccessLog/AllErrorRequestsGetsLogged",
 			"TestStackdriverAccessLog/NoClientRequestsGetsLoggedOnErrorConfigAndAllSuccessRequests",
+			"TestStackdriverAccessLog/RequestGetsLoggedAgain",
+			"TestStackdriverAccessLog/StackdriverAndAccessLogPlugin",
+			"TestStackdriverAccessLogFilter",
+			"TestStackdriverAttributeGen",
+			"TestStackdriverAuditLog",
+			"TestStackdriverCustomAccessLog",
+			"TestStackdriverGCEInstances",
+			"TestStackdriverGenericNode",
+			"TestStackdriverMetricExpiry",
+			"TestStackdriverParallel",
 			"TestStackdriverPayload",
 			"TestStackdriverPayloadGateway",
-			"TestStackdriverPayloadWithTLS",
 			"TestStackdriverPayloadUtf8",
-			"TestStackdriverReload",
-			"TestStackdriverVMReload",
-			"TestStackdriverParallel",
-			"TestStackdriverGCEInstances",
-			"TestStackdriverTCPMetadataExchange/BaseCase",
-			"TestStackdriverTCPMetadataExchange/NoAlpn",
-			"TestStackdriverAttributeGen",
-			"TestStackdriverCustomAccessLog",
+			"TestStackdriverPayloadWithTLS",
+			"TestStackdriverRbacAccessDenied/ActionAllow",
 			"TestStackdriverRbacAccessDenied/ActionBoth",
 			"TestStackdriverRbacAccessDenied/ActionDeny",
-			"TestStackdriverRbacAccessDenied/ActionAllow",
 			"TestStackdriverRbacTCPDryRun",
-			"TestStackdriverMetricExpiry",
-			"TestStatsPayload/Default/envoy.wasm.runtime.null",
-			"TestStatsPayload/Customized/envoy.wasm.runtime.null",
-			"TestStatsPayload/UseHostHeader/envoy.wasm.runtime.null",
-			"TestStatsPayload/DisableHostHeader/envoy.wasm.runtime.null",
-			"TestStatsPayload/Default/envoy.wasm.runtime.v8",
-			"TestStatsPayload/Customized/envoy.wasm.runtime.v8",
-			"TestStatsPayload/UseHostHeader/envoy.wasm.runtime.v8",
-			"TestStatsPayload/DisableHostHeader/envoy.wasm.runtime.v8",
-			"TestStackdriverGenericNode",
-			"TestStatsEndpointLabels",
-			"TestStatsParallel",
-			"TestStatsGrpc",
-			"TestTCPMetadataExchange",
-			"TestStackdriverAuditLog",
-			"TestTCPMetadataExchangeNoAlpn",
-			"TestAttributeGen",
+			"TestStackdriverRbacTCPDryRun/BaseCase",
+			"TestStackdriverRbacTCPDryRun/NoAlpn",
+			"TestStackdriverReload",
+			"TestStackdriverTCPMetadataExchange/BaseCase",
+			"TestStackdriverTCPMetadataExchange/NoAlpn",
+			"TestStackdriverVMReload",
 			"TestStats403Failure/envoy.wasm.runtime.null",
 			"TestStats403Failure/envoy.wasm.runtime.v8",
 			"TestStats403Failure/envoy.wasm.runtime.v8#01",
+			"TestStats403Failure/#00",
 			"TestStatsECDS/envoy.wasm.runtime.null",
 			"TestStatsECDS/envoy.wasm.runtime.v8",
 			"TestStatsECDS/envoy.wasm.runtime.v8#01",
-			"TestHTTPLocalRatelimit",
-			"TestStackdriverRbacTCPDryRun/BaseCase",
-			"TestStackdriverRbacTCPDryRun/NoAlpn",
+			"TestStatsECDS/#00",
+			"TestStatsEndpointLabels/envoy.wasm.runtime.null",
+			"TestStatsEndpointLabels/envoy.wasm.runtime.v8",
+			"TestStatsEndpointLabels/envoy.wasm.runtime.v8#01",
+			"TestStatsEndpointLabels/#00",
+			"TestStatsServerWaypointProxy",
+			"TestStatsGrpc/envoy.wasm.runtime.null",
+			"TestStatsGrpc/envoy.wasm.runtime.v8",
+			"TestStatsGrpc/envoy.wasm.runtime.v8#01",
+			"TestStatsGrpc/#00",
+			"TestStatsGrpcStream",
+			"TestStatsParallel/Default",
+			"TestStatsParallel/Customized",
+			"TestStatsPayload/Customized/envoy.wasm.runtime.null",
+			"TestStatsPayload/Customized/envoy.wasm.runtime.v8",
+			"TestStatsPayload/Customized/envoy.wasm.runtime.v8#01",
+			"TestStatsPayload/Customized/",
+			"TestStatsPayload/Default/envoy.wasm.runtime.null",
+			"TestStatsPayload/Default/envoy.wasm.runtime.v8",
+			"TestStatsPayload/Default/envoy.wasm.runtime.v8#01",
+			"TestStatsPayload/Default/",
+			"TestStatsPayload/DisableHostHeader/envoy.wasm.runtime.null",
+			"TestStatsPayload/DisableHostHeader/envoy.wasm.runtime.v8",
+			"TestStatsPayload/DisableHostHeader/envoy.wasm.runtime.v8#01",
+			"TestStatsPayload/DisableHostHeader/",
+			"TestStatsPayload/UseHostHeader/envoy.wasm.runtime.null",
+			"TestStatsPayload/UseHostHeader/envoy.wasm.runtime.v8",
+			"TestStatsPayload/UseHostHeader/envoy.wasm.runtime.v8#01",
+			"TestStatsPayload/UseHostHeader/",
+			"TestStatsParserRegression",
+			"TestTCPMetadataExchange/envoy.wasm.runtime.null",
+			"TestTCPMetadataExchange/#00",
+			"TestTCPMetadataExchangeNoAlpn/envoy.wasm.runtime.null",
+			"TestTCPMetadataExchangeNoAlpn/#00",
 		},
 	}
 }
