@@ -50,6 +50,9 @@ var Runtimes = []struct {
 	WasmRuntime                string
 }{
 	{
+		// native filter
+	},
+	{
 		MetadataExchangeFilterCode: "inline_string: \"envoy.wasm.metadata_exchange\"",
 		StatsFilterCode:            "inline_string: \"envoy.wasm.stats\"",
 		WasmRuntime:                "envoy.wasm.runtime.null",
@@ -83,11 +86,11 @@ var TestCases = []struct {
 		ClientConfig: "testdata/stats/client_config.yaml",
 		ServerConfig: "testdata/stats/server_config.yaml",
 		ClientStats: map[string]driver.StatMatcher{
-			"istio_requests_total": &driver.ExactStat{"testdata/metric/client_request_total.yaml.tmpl"},
+			"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_request_total.yaml.tmpl"},
 		},
 		ServerStats: map[string]driver.StatMatcher{
-			"istio_requests_total": &driver.ExactStat{"testdata/metric/server_request_total.yaml.tmpl"},
-			"istio_build":          &driver.ExactStat{"testdata/metric/istio_build.yaml"},
+			"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
+			"istio_build":          &driver.ExactStat{Metric: "testdata/metric/istio_build.yaml"},
 		},
 		TestParallel: true,
 	},
@@ -96,13 +99,13 @@ var TestCases = []struct {
 		ClientConfig: "testdata/stats/client_config_customized.yaml.tmpl",
 		ServerConfig: "testdata/stats/server_config.yaml",
 		ClientStats: map[string]driver.StatMatcher{
-			"istio_custom":                        &driver.ExactStat{"testdata/metric/client_custom_metric.yaml.tmpl"},
-			"istio_requests_total":                &driver.ExactStat{"testdata/metric/client_request_total_customized.yaml.tmpl"},
-			"istio_request_duration_milliseconds": &driver.MissingStat{"istio_request_duration_milliseconds"},
+			"istio_custom":                        &driver.ExactStat{Metric: "testdata/metric/client_custom_metric.yaml.tmpl"},
+			"istio_requests_total":                &driver.ExactStat{Metric: "testdata/metric/client_request_total_customized.yaml.tmpl"},
+			"istio_request_duration_milliseconds": &driver.MissingStat{Metric: "istio_request_duration_milliseconds"},
 		},
 		ServerStats: map[string]driver.StatMatcher{
-			"istio_requests_total": &driver.ExactStat{"testdata/metric/server_request_total.yaml.tmpl"},
-			"istio_build":          &driver.ExactStat{"testdata/metric/istio_build.yaml"},
+			"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
+			"istio_build":          &driver.ExactStat{Metric: "testdata/metric/istio_build.yaml"},
 		},
 		TestParallel: true,
 	},
@@ -112,7 +115,7 @@ var TestCases = []struct {
 		ServerConfig:      "testdata/stats/server_config.yaml",
 		ServerClusterName: "host_header",
 		ClientStats: map[string]driver.StatMatcher{
-			"istio_requests_total": &driver.ExactStat{"testdata/metric/host_header_fallback.yaml.tmpl"},
+			"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/host_header_fallback.yaml.tmpl"},
 		},
 		ServerStats: map[string]driver.StatMatcher{},
 	},
@@ -122,10 +125,10 @@ var TestCases = []struct {
 		ServerConfig:      "testdata/stats/server_config_disable_header_fallback.yaml",
 		ServerClusterName: "host_header",
 		ClientStats: map[string]driver.StatMatcher{
-			"istio_requests_total": &driver.ExactStat{"testdata/metric/client_disable_host_header_fallback.yaml.tmpl"},
+			"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_disable_host_header_fallback.yaml.tmpl"},
 		},
 		ServerStats: map[string]driver.StatMatcher{
-			"istio_requests_total": &driver.ExactStat{"testdata/metric/server_disable_host_header_fallback.yaml.tmpl"},
+			"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_disable_host_header_fallback.yaml.tmpl"},
 		},
 		ElideServerMetadata: true,
 	},
@@ -154,11 +157,14 @@ func enableStats(t *testing.T, vars map[string]string) {
 }
 
 func TestStatsPayload(t *testing.T) {
-	env.SkipTSanASan(t)
 	for _, testCase := range TestCases {
 		for _, runtime := range Runtimes {
 			t.Run(testCase.Name+"/"+runtime.WasmRuntime, func(t *testing.T) {
 				env.SkipWasm(t, runtime.WasmRuntime)
+				clientStats := make(map[string]driver.StatMatcher)
+				for metric, values := range testCase.ClientStats {
+					clientStats[metric] = values
+				}
 				params := driver.NewTestParams(t, map[string]string{
 					"RequestCount":               "10",
 					"MetadataExchangeFilterCode": runtime.MetadataExchangeFilterCode,
@@ -174,7 +180,7 @@ func TestStatsPayload(t *testing.T) {
 				params.Vars["ServerMetadata"] = params.LoadTestData("testdata/server_node_metadata.json.tmpl")
 				enableStats(t, params.Vars)
 				if err := (&driver.Scenario{
-					[]driver.Step{
+					Steps: []driver.Step{
 						&driver.XDS{},
 						&driver.Update{
 							Node:      "client",
@@ -185,7 +191,7 @@ func TestStatsPayload(t *testing.T) {
 						&driver.Update{Node: "server", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/server.yaml.tmpl")}},
 						&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/server.yaml.tmpl")},
 						&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
-						&driver.Sleep{1 * time.Second},
+						&driver.Sleep{Duration: 1 * time.Second},
 						&driver.Repeat{
 							N: 10,
 							Step: &driver.HTTPCall{
@@ -193,8 +199,8 @@ func TestStatsPayload(t *testing.T) {
 								Body: "hello, world!",
 							},
 						},
-						&driver.Stats{params.Ports.ClientAdmin, testCase.ClientStats},
-						&driver.Stats{params.Ports.ServerAdmin, testCase.ServerStats},
+						&driver.Stats{AdminPort: params.Ports.ClientAdmin, Matchers: clientStats},
+						&driver.Stats{AdminPort: params.Ports.ServerAdmin, Matchers: testCase.ServerStats},
 					},
 				}).Run(params); err != nil {
 					t.Fatal(err)
@@ -233,18 +239,18 @@ func TestStatsParallel(t *testing.T) {
 			serverListenerTemplate := driver.LoadTestData("testdata/listener/server.yaml.tmpl")
 
 			if err := (&driver.Scenario{
-				[]driver.Step{
+				Steps: []driver.Step{
 					&driver.XDS{},
 					&driver.Update{Node: "client", Version: "0", Listeners: []string{clientListenerTemplate}},
 					&driver.Update{Node: "server", Version: "0", Listeners: []string{serverListenerTemplate}},
 					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/server.yaml.tmpl")},
 					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
-					&driver.Sleep{1 * time.Second},
+					&driver.Sleep{Duration: 1 * time.Second},
 					driver.Get(params.Ports.ClientPort, "hello, world!"),
 					&driver.Fork{
 						Fore: &driver.Scenario{
-							[]driver.Step{
-								&driver.Sleep{1 * time.Second},
+							Steps: []driver.Step{
+								&driver.Sleep{Duration: 1 * time.Second},
 								&driver.Repeat{
 									Duration: 9 * time.Second,
 									Step:     driver.Get(params.Ports.ClientPort, "hello, world!"),
@@ -255,7 +261,7 @@ func TestStatsParallel(t *testing.T) {
 						Back: &driver.Repeat{
 							Duration: 10 * time.Second,
 							Step: &driver.Scenario{
-								[]driver.Step{
+								Steps: []driver.Step{
 									&driver.Update{
 										Node:      "client",
 										Version:   "{{.N}}",
@@ -267,13 +273,13 @@ func TestStatsParallel(t *testing.T) {
 										Listeners: []string{serverListenerTemplate},
 									},
 									// may need short delay so we don't eat all the CPU
-									&driver.Sleep{100 * time.Millisecond},
+									&driver.Sleep{Duration: 100 * time.Millisecond},
 								},
 							},
 						},
 					},
-					&driver.Stats{params.Ports.ClientAdmin, testCase.ClientStats},
-					&driver.Stats{params.Ports.ServerAdmin, testCase.ServerStats},
+					&driver.Stats{AdminPort: params.Ports.ClientAdmin, Matchers: testCase.ClientStats},
+					&driver.Stats{AdminPort: params.Ports.ServerAdmin, Matchers: testCase.ServerStats},
 				},
 			}).Run(params); err != nil {
 				t.Fatal(err)
@@ -283,50 +289,55 @@ func TestStatsParallel(t *testing.T) {
 }
 
 func TestStatsGrpc(t *testing.T) {
-	params := driver.NewTestParams(t, map[string]string{
-		"RequestCount":               "10",
-		"MetadataExchangeFilterCode": "inline_string: \"envoy.wasm.metadata_exchange\"",
-		"StatsFilterCode":            "inline_string: \"envoy.wasm.stats\"",
-		"WasmRuntime":                "envoy.wasm.runtime.null",
-		"DisableDirectResponse":      "true",
-		"UsingGrpcBackend":           "true",
-		"GrpcResponseStatus":         "7",
-		"StatsConfig":                driver.LoadTestData("testdata/bootstrap/stats.yaml.tmpl"),
-		"StatsFilterClientConfig":    driver.LoadTestJSON("testdata/stats/client_config.yaml"),
-		"StatsFilterServerConfig":    driver.LoadTestJSON("testdata/stats/server_config.yaml"),
-	}, envoye2e.ProxyE2ETests)
-	params.Vars["ClientMetadata"] = params.LoadTestData("testdata/client_node_metadata.json.tmpl")
-	params.Vars["ServerMetadata"] = params.LoadTestData("testdata/server_node_metadata.json.tmpl")
-	enableStats(t, params.Vars)
+	for _, runtime := range Runtimes {
+		t.Run(runtime.WasmRuntime, func(t *testing.T) {
+			env.SkipWasm(t, runtime.WasmRuntime)
+			params := driver.NewTestParams(t, map[string]string{
+				"RequestCount":               "10",
+				"MetadataExchangeFilterCode": runtime.MetadataExchangeFilterCode,
+				"StatsFilterCode":            runtime.StatsFilterCode,
+				"WasmRuntime":                runtime.WasmRuntime,
+				"DisableDirectResponse":      "true",
+				"UsingGrpcBackend":           "true",
+				"GrpcResponseStatus":         "7",
+				"StatsConfig":                driver.LoadTestData("testdata/bootstrap/stats.yaml.tmpl"),
+				"StatsFilterClientConfig":    driver.LoadTestJSON("testdata/stats/client_config.yaml"),
+				"StatsFilterServerConfig":    driver.LoadTestJSON("testdata/stats/server_config.yaml"),
+			}, envoye2e.ProxyE2ETests)
+			params.Vars["ClientMetadata"] = params.LoadTestData("testdata/client_node_metadata.json.tmpl")
+			params.Vars["ServerMetadata"] = params.LoadTestData("testdata/server_node_metadata.json.tmpl")
+			enableStats(t, params.Vars)
 
-	if err := (&driver.Scenario{
-		Steps: []driver.Step{
-			&driver.XDS{},
-			&driver.Update{Node: "client", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/client.yaml.tmpl")}},
-			&driver.Update{Node: "server", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/server.yaml.tmpl")}},
-			&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/server.yaml.tmpl")},
-			&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
-			&driver.Sleep{Duration: 1 * time.Second},
-			&driver.GrpcServer{},
-			&driver.GrpcCall{
-				ReqCount:   10,
-				WantStatus: status.New(codes.PermissionDenied, "denied"),
-			},
-			&driver.Stats{
-				AdminPort: params.Ports.ServerAdmin,
-				Matchers: map[string]driver.StatMatcher{
-					"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
+			if err := (&driver.Scenario{
+				Steps: []driver.Step{
+					&driver.XDS{},
+					&driver.Update{Node: "client", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/client.yaml.tmpl")}},
+					&driver.Update{Node: "server", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/server.yaml.tmpl")}},
+					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/server.yaml.tmpl")},
+					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
+					&driver.Sleep{Duration: 1 * time.Second},
+					&driver.GrpcServer{},
+					&driver.GrpcCall{
+						ReqCount:   10,
+						WantStatus: status.New(codes.PermissionDenied, "denied"),
+					},
+					&driver.Stats{
+						AdminPort: params.Ports.ServerAdmin,
+						Matchers: map[string]driver.StatMatcher{
+							"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
+						},
+					},
+					&driver.Stats{
+						AdminPort: params.Ports.ClientAdmin,
+						Matchers: map[string]driver.StatMatcher{
+							"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_request_total.yaml.tmpl"},
+						},
+					},
 				},
-			},
-			&driver.Stats{
-				AdminPort: params.Ports.ClientAdmin,
-				Matchers: map[string]driver.StatMatcher{
-					"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_request_total.yaml.tmpl"},
-				},
-			},
-		},
-	}).Run(params); err != nil {
-		t.Fatal(err)
+			}).Run(params); err != nil {
+				t.Fatal(err)
+			}
+		})
 	}
 }
 
@@ -414,11 +425,8 @@ func TestAttributeGen(t *testing.T) {
 			params := driver.NewTestParams(t, map[string]string{
 				"RequestCount":               "10",
 				"MetadataExchangeFilterCode": "inline_string: \"envoy.wasm.metadata_exchange\"",
-				"StatsFilterCode":            "inline_string: \"envoy.wasm.stats\"",
 				"AttributeGenFilterConfig":   runtime.AttributeGenFilterCode,
 				"AttributeGenWasmRuntime":    runtime.WasmRuntime,
-				"WasmRuntime":                "envoy.wasm.runtime.null",
-				"StatsConfig":                driver.LoadTestData("testdata/bootstrap/stats.yaml.tmpl"),
 				"StatsFilterClientConfig":    driver.LoadTestJSON("testdata/stats/client_config.yaml"),
 				"StatsFilterServerConfig":    driver.LoadTestJSON("testdata/stats/request_classification_config.yaml"),
 				"ResponseCodeClass":          "2xx",
@@ -429,7 +437,7 @@ func TestAttributeGen(t *testing.T) {
 			params.Vars["ServerHTTPFilters"] = driver.LoadTestData("testdata/filters/attributegen.yaml.tmpl") + "\n" +
 				params.Vars["ServerHTTPFilters"]
 			if err := (&driver.Scenario{
-				[]driver.Step{
+				Steps: []driver.Step{
 					&driver.XDS{},
 					&driver.Update{Node: "client", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/client.yaml.tmpl")}},
 					&driver.Update{Node: "server", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/server.yaml.tmpl")}},
@@ -443,8 +451,8 @@ func TestAttributeGen(t *testing.T) {
 							Body: "hello, world!",
 						},
 					},
-					&driver.Stats{params.Ports.ServerAdmin, map[string]driver.StatMatcher{
-						"istio_requests_total": &driver.ExactStat{"testdata/metric/server_request_total.yaml.tmpl"},
+					&driver.Stats{AdminPort: params.Ports.ServerAdmin, Matchers: map[string]driver.StatMatcher{
+						"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
 					}},
 				},
 			}).Run(params); err != nil {
@@ -467,7 +475,7 @@ func TestStatsParserRegression(t *testing.T) {
 	params.Vars["StatsFilterClientConfig"] = driver.LoadTestJSON("testdata/stats/client_config_customized.yaml.tmpl")
 	listener1 := params.LoadTestData("testdata/listener/client.yaml.tmpl")
 	if err := (&driver.Scenario{
-		[]driver.Step{
+		Steps: []driver.Step{
 			&driver.XDS{},
 			&driver.Update{
 				Node:      "client",
@@ -476,14 +484,14 @@ func TestStatsParserRegression(t *testing.T) {
 				Listeners: []string{listener0},
 			},
 			&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
-			&driver.Sleep{1 * time.Second},
+			&driver.Sleep{Duration: 1 * time.Second},
 			&driver.Update{
 				Node:      "client",
 				Version:   "1",
 				Clusters:  []string{params.LoadTestData("testdata/cluster/server.yaml.tmpl")},
 				Listeners: []string{listener1},
 			},
-			&driver.Sleep{1 * time.Second},
+			&driver.Sleep{Duration: 1 * time.Second},
 		},
 	}).Run(params); err != nil {
 		t.Fatal(err)
@@ -513,7 +521,7 @@ func TestStats403Failure(t *testing.T) {
 				params.LoadTestData("testdata/filters/rbac.yaml.tmpl") + "\n" +
 				driver.LoadTestData("testdata/filters/stats_inbound.yaml.tmpl")
 			if err := (&driver.Scenario{
-				[]driver.Step{
+				Steps: []driver.Step{
 					&driver.XDS{},
 					&driver.Update{Node: "client", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/client.yaml.tmpl")}},
 					&driver.Update{Node: "server", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/server.yaml.tmpl")}},
@@ -528,8 +536,8 @@ func TestStats403Failure(t *testing.T) {
 							ResponseCode: 403,
 						},
 					},
-					&driver.Stats{params.Ports.ServerAdmin, map[string]driver.StatMatcher{
-						"istio_requests_total": &driver.ExactStat{"testdata/metric/server_request_total.yaml.tmpl"},
+					&driver.Stats{AdminPort: params.Ports.ServerAdmin, Matchers: map[string]driver.StatMatcher{
+						"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
 					}},
 				},
 			}).Run(params); err != nil {
@@ -567,7 +575,7 @@ func TestStatsECDS(t *testing.T) {
 				},
 			}
 			if err := (&driver.Scenario{
-				[]driver.Step{
+				Steps: []driver.Step{
 					&driver.XDS{},
 					&driver.Update{
 						Node:      "client",
@@ -579,7 +587,7 @@ func TestStatsECDS(t *testing.T) {
 					updateExtensions,
 					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/server.yaml.tmpl")},
 					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
-					&driver.Sleep{1 * time.Second},
+					&driver.Sleep{Duration: 1 * time.Second},
 					&driver.Repeat{
 						N: 10,
 						Step: &driver.HTTPCall{
@@ -587,11 +595,11 @@ func TestStatsECDS(t *testing.T) {
 							Body: "hello, world!",
 						},
 					},
-					&driver.Stats{params.Ports.ClientAdmin, map[string]driver.StatMatcher{
-						"istio_requests_total": &driver.ExactStat{"testdata/metric/client_request_total.yaml.tmpl"},
+					&driver.Stats{AdminPort: params.Ports.ClientAdmin, Matchers: map[string]driver.StatMatcher{
+						"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_request_total.yaml.tmpl"},
 					}},
-					&driver.Stats{params.Ports.ServerAdmin, map[string]driver.StatMatcher{
-						"istio_requests_total": &driver.ExactStat{"testdata/metric/server_request_total.yaml.tmpl"},
+					&driver.Stats{AdminPort: params.Ports.ServerAdmin, Matchers: map[string]driver.StatMatcher{
+						"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_request_total.yaml.tmpl"},
 					}},
 				},
 			}).Run(params); err != nil {
@@ -602,6 +610,51 @@ func TestStatsECDS(t *testing.T) {
 }
 
 func TestStatsEndpointLabels(t *testing.T) {
+	for _, runtime := range Runtimes {
+		t.Run(runtime.WasmRuntime, func(t *testing.T) {
+			env.SkipWasm(t, runtime.WasmRuntime)
+			params := driver.NewTestParams(t, map[string]string{
+				"RequestCount":               "10",
+				"MetadataExchangeFilterCode": runtime.MetadataExchangeFilterCode,
+				"StatsFilterCode":            runtime.StatsFilterCode,
+				"WasmRuntime":                runtime.WasmRuntime,
+				"EnableEndpointMetadata":     "true",
+				"StatsConfig":                driver.LoadTestData("testdata/bootstrap/stats.yaml.tmpl"),
+				"StatsFilterClientConfig":    driver.LoadTestJSON("testdata/stats/client_config.yaml"),
+			}, envoye2e.ProxyE2ETests)
+			params.Vars["ClientMetadata"] = params.LoadTestData("testdata/client_node_metadata.json.tmpl")
+			params.Vars["ServerMetadata"] = params.LoadTestData("testdata/server_node_metadata.json.tmpl")
+			params.Vars["ClientHTTPFilters"] = driver.LoadTestData("testdata/filters/stats_outbound.yaml.tmpl")
+			if err := (&driver.Scenario{
+				Steps: []driver.Step{
+					&driver.XDS{},
+					&driver.Update{Node: "client", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/client.yaml.tmpl")}},
+					&driver.Update{Node: "server", Version: "0", Listeners: []string{params.LoadTestData("testdata/listener/server.yaml.tmpl")}},
+					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/client.yaml.tmpl")},
+					&driver.Envoy{Bootstrap: params.LoadTestData("testdata/bootstrap/server.yaml.tmpl")},
+					&driver.Sleep{Duration: 1 * time.Second},
+					&driver.Repeat{
+						N: 10,
+						Step: &driver.HTTPCall{
+							Port:         params.Ports.ClientPort,
+							ResponseCode: 200,
+						},
+					},
+					&driver.Stats{
+						AdminPort: params.Ports.ClientAdmin,
+						Matchers: map[string]driver.StatMatcher{
+							"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_request_total_endpoint_labels.yaml.tmpl"},
+						},
+					},
+				},
+			}).Run(params); err != nil {
+				t.Fatal(err)
+			}
+		})
+	}
+}
+
+func TestStatsServerWaypointProxy(t *testing.T) {
 	params := driver.NewTestParams(t, map[string]string{
 		"RequestCount":               "10",
 		"MetadataExchangeFilterCode": "inline_string: \"envoy.wasm.metadata_exchange\"",
@@ -609,11 +662,14 @@ func TestStatsEndpointLabels(t *testing.T) {
 		"WasmRuntime":                "envoy.wasm.runtime.null",
 		"EnableEndpointMetadata":     "true",
 		"StatsConfig":                driver.LoadTestData("testdata/bootstrap/stats.yaml.tmpl"),
-		"StatsFilterClientConfig":    driver.LoadTestJSON("testdata/stats/client_config.yaml"),
+		"StatsFilterServerConfig":    driver.LoadTestJSON("testdata/stats/server_waypoint_proxy_config.yaml"),
 	}, envoye2e.ProxyE2ETests)
 	params.Vars["ClientMetadata"] = params.LoadTestData("testdata/client_node_metadata.json.tmpl")
-	params.Vars["ServerMetadata"] = params.LoadTestData("testdata/server_node_metadata.json.tmpl")
-	params.Vars["ClientHTTPFilters"] = driver.LoadTestData("testdata/filters/stats_outbound.yaml.tmpl")
+	params.Vars["ServerMetadata"] = params.LoadTestData("testdata/server_waypoint_proxy_node_metadata.json.tmpl")
+	params.Vars["ServerHTTPFilters"] = driver.LoadTestData("testdata/filters/mx_inbound.yaml.tmpl") + "\n" +
+		driver.LoadTestData("testdata/filters/stats_inbound.yaml.tmpl")
+	params.Vars["ClientHTTPFilters"] = driver.LoadTestData("testdata/filters/mx_outbound.yaml.tmpl")
+
 	if err := (&driver.Scenario{
 		Steps: []driver.Step{
 			&driver.XDS{},
@@ -630,9 +686,9 @@ func TestStatsEndpointLabels(t *testing.T) {
 				},
 			},
 			&driver.Stats{
-				AdminPort: params.Ports.ClientAdmin,
+				AdminPort: params.Ports.ServerAdmin,
 				Matchers: map[string]driver.StatMatcher{
-					"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/client_request_total_endpoint_labels.yaml.tmpl"},
+					"istio_requests_total": &driver.ExactStat{Metric: "testdata/metric/server_waypoint_proxy_request_total.yaml.tmpl"},
 				},
 			},
 		},
